requirements.txt
accelerate               
ctransformers            
datasets                
faiss-cpu                
fsspec
gpt4all
huggingface-hub 
langchain
langchain-community
langchain-core
langchain-openai
langchain-pinecone
langchain-tests
langchain-text-splitters
langsmith
numpy
openai
pandas
pinecone
pinecone-plugin-interface
pydantic 
pydantic_core
pydantic-settings
pypdf
python-dotenv
requests
scikit-learn
sentence-transformers
tokenizers
torch
tqdm
transformers
typing_extensions
typing-inspect   
rank-llm
rank_bm25
langserve 
fastapi 
uvicorn
langsmith